requirements.txt
cmake==3.29.2
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
dlib==19.18.0
face_recognition==1.3.0